Reorganize dart files
