+++ lib/list_page.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:idea_board/idea_card.dart';
-import 'package:idea_board/ideas.dart';
+import 'package:idea_board/legacy/ideas.dart';
 import 'package:provider/provider.dart';
+import 'package:idea_board/model/idea.dart';
 
 class ListPage extends StatelessWidget {
   const ListPage({Key? key}) : super(key: key);
